Simplify workflow to run unit and integration tests together
This is needed because of our php requirements and phpunit version of the shop. Coverage merging on php 8.1 doesnt work with phpunit 9. Also, unit tests are running in several seconds, installing separate run just for those seconds is not adequate. Locally it still can be run separately for fast check, but also running all integrations doesnt take long as well.

Signed-off-by: Anton Fedurtsya <anton@fedurtsya.com>